refactor: move responsibility for dispatcher flag to Nav component
