+++ assets/src/components/nav.tsx
@@ -1,6 +1,7 @@
 import React from "react"
 import { OpenView } from "../state"
 import TabBar from "./tabBar"
+import appData from "../appData"
 
 interface Props {
   pickerContainerIsVisible: boolean
@@ -17,8 +18,18 @@ export const Nav: React.FC<Props> = ({
       <TabBar
         pickerContainerIsVisible={pickerContainerIsVisible}
         openView={openView}
+        dispatcherFlag={readDispatcherFlag()}
       />
       {children}
     </>
   )
+}
+
+const readDispatcherFlag = (): boolean => {
+  const data = appData()
+  if (!data) {
+    return false
+  }
+
+  return data.dispatcherFlag === "true"
 }
